Update ximtool start screen
To avoid the need to re-create the start screen, the version number is
removed there. However, we still keep the NOAO logo. Since the ximtool
code is almost unchanged, it is probably better to keep it as a kind
of acknowledgement.

We also replace the mkhdr.c tool by a Python script that directly reads the logo FITS file